Final Changes for the session 2.0

+++ AutomatedTicketSystemProject_Group6/AutomatedTicketSystemProject_Group6/CustomerInterface.cs
@@ -19,9 +19,9 @@ namespace AutomatedTicketSystemProject_Group6
 
         private void btnSave_Click(object sender, EventArgs e)
         {
-            int cardNo;
+            float cardNo;
             int CVV;
-            if (int.TryParse(txtCardNo.Text, out cardNo))
+            if (float.TryParse(txtCardNo.Text, out cardNo))
             {
                 if (txtCardNo.Text.Length != 16)
                 {
